Use FEN to generate diagram file names
Before this commit, the move list was used to generate the name of
diagrams. For longer variations, this caused problems as long file names
are not supported by the operating system. Now, we use FEN to generate
file names, replacing `/` with `\` to avoid issues with UNIX file
system. However this change causes the same issues with Windows, which
I don't really care about.

Fixes https://github.com/khanalizadeh/automate-boring-stuff/issues/1